accounts/holding: Rename module/calls
Move accounts over to "holdings" in preparation for getting holdings at
different granularity levels. The math is the ~same but grouping may be
different for different use cases. A list of asset holdings may be for
an integraiton, provider, or global.

Once a list of all holdings can be found then that may provide a good
view into setting the mapping of holding -> asset, then price